python/sentinel-agents/pyproject.toml
[project]
name = "sentinel-agents"
version = "0.1.0"
description = "Sentinel Agent Framework — autonomous LLM-powered cyber defense agents"
requires-python = ">=3.11"
dependencies = [
    "sentinel-api",
    "sentinel-policy",
    "anthropic>=0.40.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
]

[tool.uv.sources]
sentinel-api = { workspace = true }
sentinel-policy = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentinel_agents"]
